refactor: Changed blocks.py to block.py for consistency.
- Changed how `Block` is initialized.
- Changed `load_blocks_from_schema` into a `@classmethod` to simplify imports.
